split rewrite out into files
